IS 1182 separate folder for dkg files
