--- core/schains/dkg/__init__.py
@@ -1,24 +1,0 @@
-#   -*- coding: utf-8 -*-
-#
-#   This file is part of SKALE Admin
-#
-#   Copyright (C) 2021 SKALE Labs
-#
-#   This program is free software: you can redistribute it and/or modify
-#   it under the terms of the GNU Affero General Public License as published by
-#   the Free Software Foundation, either version 3 of the License, or
-#   (at your option) any later version.
-#
-#   This program is distributed in the hope that it will be useful,
-#   but WITHOUT ANY WARRANTY; without even the implied warranty of
-#   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#   GNU Affero General Public License for more details.
-#
-#   You should have received a copy of the GNU Affero General Public License
-#   along with this program.  If not, see <https://www.gnu.org/licenses/>.
-
-# ruff: noqa: F401
-
-from core.schains.dkg.main import get_dkg_client, run_dkg, save_dkg_results
-from core.schains.dkg.structures import DKGStatus, DKGStep
-from core.schains.dkg.utils import DkgError, get_secret_key_share_filepath
